Moved popuplink component inside map component
