+++ src/app/components/app.module.ts
@@ -65,7 +65,7 @@ import { SidebarFooterComponent } from './sidebar/sidebar-footer/sidebar-footer.
 import { SidebarItemChildComponent } from './sidebar/sidebar-menu/sidebar-item/sidebar-item-child/sidebar-item-child.component';
 import { ToolsComponent } from './map/tools/tools.component';
 import { VisibleLayersComponent } from './map/visible-layers/visible-layers.component';
-import { PopupLinkComponent } from './popup-link/popup-link.component';
+import { PopupLinkComponent } from './map/popup-link/popup-link.component';
 
 registerLocaleData(localePt, 'pt');
