+++ src/app/components/map/popup-link/popup-link.component.ts
